+++ RayTracingInAWeekend/hittable.h
@@ -12,7 +12,7 @@ public:
     vec3 p;
     vec3 normal;
     std::shared_ptr<material> mat;
-    double t;
+    float t;
     bool front_face;
 
     void set_face_normal(const ray& r, const vec3& outward_normal)
